Add files via upload
The code should work in theory, but in practice rdrobust  and rdplot command takes ages and i never managed to get any result (people on the internet say that it's some kind of bug with the function itself).  So the main output of the code is latex code for the table containing the regression results (excluding the last panel again because of rdrobust issue).